Added overlay layers for each connector type.
All connector types are fetched by looping through the json response and an javascript object is created to store each type with a L.LayerGroup().
Added a zoom-in and center you current location function when user clike the button "Find Stations Near Me".
Due to the global and local variable and its scope, we need to do everything in to d3.json() response as there is no way to call any variable outside the response, which in this case is the created object overlaysDict.